requirements.txt
boto3==1.9.71
python-dotenv==0.10.3
Flask==1.1.2
